test: persist single-stream e2e outputs
