+++ examples/conftest.py
@@ -1,1 +1,1 @@
-from tests.conftest import *  # noqa: F401,F403
+from tests.utils.pytest_fixtures import *  # noqa: F401,F403
