+++ examples/object-detection/yolo26-object-detector/common/config.yaml
@@ -5,7 +5,7 @@ io:
   input_dir: assets/test_images
   output_dir: sandbox/yolo26-object-detector
 decode:
-  score_threshold: 0.25
+  score_threshold: 0.65
   nms_iou: 0.45
   max_detections: 100
 runtime:
